ohh yahhh
what's done on the audio engine side:

* audio pipeline structure (moved to float msecs versus ints, still needs to be tested a bit "live" as it might have issues of partial frames)
* mmap'd wave file playing
* pitch shifting (seems to take like no CPU)
* audio compositing (two turntables mixed)

what needs being done:

* two-way buffering
* scratching-mode (needs two-way buffer)
* multi-threaded file reading/buffering
* parameter optimization for playback buffers, optimum file read/cache chunk speeds
* mp3-based forward playing (should be easy once the above is implemented since we already have libmad working)
* mp3 seeking (hard, but we can wait out on this feature)
* xfader/volume implementation, using an adjustable sigmoid function for the cut-off ramp (especially needed for normal djing versus scratching mode)
* EQ (noah should easily implement)

+++ AudioCompositor.h
@@ -12,15 +12,15 @@
 @interface AudioCompositor : NSObject <AudioSource> 
 {
 	AUDIO_SHORTS_PTR buffer;
-	int bufferSizeInMsec;
-	int bufferSizeInBytes;
+	float bufferSizeInMsec;
+	unsigned int bufferSizeInBytes;
 
 	id <AudioSource> sourceA;
 	id <AudioSource> sourceB;
 	float xFader;
 }
 
-- (AUDIO_SHORTS_PTR) getAudio:(int) msec;
+- (AUDIO_SHORTS_PTR) getAudio:(float) msec;
 
 // can only set the sources on creation for thread performance reasons
 - (id) initWithSourceA:(id <AudioSource>) a andSourceB:(id <AudioSource>) b;
