+++ AudioNoiseSource.h
@@ -11,10 +11,10 @@
 @interface AudioNoiseSource : NSObject <AudioSource>
 {
 	AUDIO_SHORTS_PTR buffer;
-	int bufferSizeInMsec;
-	int bufferSizeInBytes;
+	float bufferSizeInMsec;
+	unsigned int bufferSizeInBytes;
 }
 
-- (AUDIO_SHORTS_PTR) getAudio:(int) msec;
+- (AUDIO_SHORTS_PTR) getAudio:(float) msec;
 
 @end
